[Refactor] - Decomposed Layout and LayoutRoute into 2  separated files
